+++ src/NEFDashboard/containers/signUp/index.jsx
@@ -1,5 +1,15 @@
 import React, { useContext, useState, useEffect } from "react";
-import { Select, Row, Col, Flex, Form, Input, Button, Divider } from "antd";
+import {
+  Select,
+  Row,
+  Col,
+  Flex,
+  Form,
+  Input,
+  Button,
+  Divider,
+  Checkbox,
+} from "antd";
 import Logo from "../../../assets/logo/logo.svg";
 import { useNavigate, useLocation } from "react-router-dom";
 import backend_API from "../../../api/backendAPI";
@@ -341,7 +351,7 @@ const SignForm = () => {
                                 {
                                   required: true,
                                   message:
-                                    "Please input your confirm password!",
+                                    "Please input your password confirmation!",
                                 },
                               ]}
                             >
@@ -447,7 +457,7 @@ const SignForm = () => {
                               rules={[
                                 {
                                   required: role === "Business",
-                                  message: "Please input your Country",
+                                  message: "Please input your country",
                                 },
                               ]}
                             >
@@ -474,15 +484,34 @@ const SignForm = () => {
                             </Form.Item>
                           </>
                         )}
-                        <Form.Item>
-                          {/* {role === "Business" && (
-                        <div
-                          className="forgot-password"
-                          onClick={() => navigate("/forgot-password")}
+                        <Form.Item
+                          name={"privacy"}
+                          valuePropName="checked"
+                          rules={[
+                            {
+                              required: true,
+                              validator: (_, value) =>
+                                value
+                                  ? Promise.resolve()
+                                  : Promise.reject(
+                                      "Please read and agree terms of use",
+                                    ),
+                            },
+                          ]}
                         >
-                          Forgot password?
-                        </div>
-                      )} */}
+                          <Checkbox className="privacy-policy-checkbox">
+                            <div className="default-text-gray privacy-policy-text">
+                              <span>I have read and agree to the </span>
+                              <a
+                                href="http://localhost:5173/login"
+                                className="privay-policy"
+                              >
+                                Terms of use
+                              </a>
+                            </div>
+                          </Checkbox>
+                        </Form.Item>
+                        <Form.Item>
                           <Button
                             type="primary"
                             htmlType="submit"